%
\name{bs.plot.quantiles}
\alias{bs.plot.quantiles}
\title{
Quantile values of Birnbaum-Saunders distribution
}
\description{
Quantiles for the goodness of fit test for the Birnbaum-Saunders distribution
from the probability plot. They are obtained
from the sample correlation from the Birnbaum-Saunders probability plot.
Monte Carlo iteration is 1E08.
}
\usage{bs.plot.quantiles}
\format{
  This data frame contains 998 rows and 1001 columns.
}
\value{Quantile values between zero and one.}
\references{ 
Park, C. and M. Wang (2023).
A goodness-of-fit test for the Birnbaum-Saunders distribution based on the probability plot.
\emph{ArXiv e-prints}, 2308.10150. \cr
\doi{10.48550/arXiv.2308.10150}
}
\author{Chanseok Park}
\keyword{datasets}
