%
\name{print.bs.test.critical.value}
\alias{print.bs.test.critical.value}
\title{
Print the critical value for bs.test
}
\description{
  Printing objects of class "bs.test.critical.value". 
}
\value{No return value, printing objects of class}
\seealso{
bs.test.critical, \code{\link{print}}
}
\keyword{print}
