%
\name{bs.plot.quantiles}
\alias{bs.plot.quantiles}
\title{
Quantile values of Birnbaum-Saunders distribution
}
\description{
Quantiles for the goodness of fit test for the Birnbaum-Saunders distribution
from the probability plot. They are obtained
from the sample correlation from the Birnbaum-Saunders probability plot.
Monte Carlo iteration is 1E08.
}
\usage{bs.plot.quantiles}
\format{
  This data frame contains 998 rows and 1001 columns.
}
\keyword{datasets}
